BACKEND/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.23.0
sqlmodel>=0.0.22
SQLAlchemy>=2.0.29
pydantic>=2.7.0
python-dotenv>=1.0.0
